Rupesh Falke
Extended description text box :

When squashing and merging a pull request (PR), the extended description text box is automatically populated with all the commit messages in the PR
. 
Commit messages:

When writing a commit message, it's recommended to write the summary line and description in the imperative mood, as if you were giving a command. For example, start the summary with "Fix", "Add", or "Change". 
Organizing information
GitHub allows you to organize information with collapsed sections. 

Relative links:

You can define relative links and image paths in your rendered files to help readers navigate to other files in your repository. GitHub will automatically transform the relative link or image path based on the branch you're currently on. 

Committing:

Committing is the act of creating a snapshot of your local repository. You can create commits with the git commit command.